+++ Inafocam.core/Modelos/TracingStudyPlanDevelopment.cs
@@ -13,7 +13,7 @@ namespace Inafocam.core.Modelos
         public long? HigherTitleSupportMatterId { get; set; }
         public double? ScoreAverageStudents { get; set; }
         public int? ApprovedStudentsQuantity { get; set; }
-        public int? Creditos { get; set; }
+        public long? Creditos { get; set; }
         public int? ReprovedStudentsQuantity { get; set; }
         public int? DesertedStudentsQuantity { get; set; }
         public short? HigherTitleMatchAssignedMatter { get; set; }
